+++ gradle/libs.versions.toml
@@ -1,5 +1,5 @@
 [versions]
-felis-dam = "1.8.2-alpha"
+felis-dam = "1.9.0-alpha"
 asm = "9.7"
 tomlkt = "0.3.7"
 clikt = "4.3.0"
