+++ src/main/kotlin/felis/language/LanguageAdapter.kt
@@ -1,0 +1,5 @@
+package felis.language
+
+interface LanguageAdapter {
+    fun <T> createInstance(specifier: String): Result<T>
+}
