+++ src/main/kotlin/felis/launcher/GameLauncher.kt
@@ -1,0 +1,5 @@
+package felis.launcher
+
+fun interface GameLauncher {
+    fun instantiate(args: Array<String>): GameInstance
+}
