+++ src/main/kotlin/felis/launcher/GameNotFound.kt
@@ -1,0 +1,3 @@
+package felis.launcher
+
+class GameNotFound(name: String) : Exception("Game '$name' could not be found in the classpath")
