+++ src/main/kotlin/felis/side/OnlyIn.kt
@@ -1,9 +1,5 @@
 package felis.side
 
-enum class Side {
-    CLIENT,
-    SERVER
-}
-
 @Target(AnnotationTarget.CLASS, AnnotationTarget.FIELD, AnnotationTarget.FUNCTION)
+@Retention(AnnotationRetention.RUNTIME)
 annotation class OnlyIn(val side: Side)
