+++ src/main/kotlin/felis/side/Side.kt
@@ -1,0 +1,6 @@
+package felis.side
+
+enum class Side {
+    CLIENT,
+    SERVER
+}
