+++ src/main/kotlin/felis/transformer/Transformation.kt
@@ -1,0 +1,6 @@
+package felis.transformer
+
+fun interface Transformation {
+    fun transform(container: ClassContainer)
+    data class Named(val name: String, val del: Transformation) : Transformation by del
+}
